+++ mitra_utils/src/lib.rs
@@ -26,5 +26,6 @@ pub mod multibase;
 pub mod multicodec;
 pub mod passwords;
 pub mod random;
+pub mod sysinfo;
 pub mod unicode;
 pub mod urls;