+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Added
+
+- Write amount of disk space available for media to log at startup.
+
 ### Changed
 
 - Do not remove posts by muted users from threads.